now extends View abstract class
