+++ src/GraphicInterface/Views/ElectivesView.java
@@ -1,22 +1,12 @@
 package src.GraphicInterface.Views;
 
-import java.io.IOException;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
 import javafx.stage.Stage;
 
-public class ElectivesView {
-	
-	private Stage stage;
+public class ElectivesView extends View{
 	private static ElectivesView instance;
 	
 	private ElectivesView(Stage stage){
-		this.stage = stage;
-	}
-	
-	public Stage getStage() {
-		return stage;
+		super(stage);
 	}
 	
 	public static ElectivesView getInstance(Stage stage){
@@ -25,16 +15,5 @@ public class ElectivesView {
         }
         return instance;
     }
-	
-	public void openStage() throws IOException {
-		Parent root = FXMLLoader.load(getClass().getResource("/src/FXML/electives.fxml"));
-        Scene scene = new Scene(root);
-        stage.setScene(scene);
-        stage.show();
-	}
-	
-	public void closeStage() {
-		stage.close();
-	}
 }
