+++ src/GraphicInterface/Views/SelectionView.java
@@ -1,23 +1,12 @@
 package src.GraphicInterface.Views;
 
-import java.io.IOException;
-
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
 import javafx.stage.Stage;
 
-public class SelectionView {
-	
-	private Stage stage;
+public class SelectionView extends View{
 	private static SelectionView instance;
 	
 	private SelectionView(Stage stage){
-		this.stage = stage;
-	}
-	
-	public Stage getStage() {
-		return stage;
+		super(stage);
 	}
 	
 	public static SelectionView getInstance(Stage stage){
@@ -26,15 +15,4 @@ public class SelectionView {
         }
         return instance;
     }
-	
-	public void openStage() throws IOException {
-		Parent root = FXMLLoader.load(getClass().getResource("/src/FXML/selection.fxml"));
-        Scene scene = new Scene(root);
-        stage.setScene(scene);
-        stage.show();
-	}
-	
-	public void closeStage() {
-		stage.close();
-	}
 }
